# Kwkim5 split performance logger