--- a/LgSwArchCode2021Distv2/LgFaceRecDemoTCP_Jetson_Nano/src/NetworkInterface.h
+++ b/LgSwArchCode2021Distv2/LgFaceRecDemoTCP_Jetson_Nano/src/NetworkInterface.h
@@ -31,6 +31,7 @@ public:
     void PushDataToSend(cv::Mat &data);
     void Stop();
     void StopWhenEmpty();
+    size_t GetCurrentTransmitQueueSize();
 
 private:
     NetworkInterface();