thread 가 분리되니까, 로그가 잘 안 보여서 별도 파일로 분리해서 같이 출력되게 수정했습니다.
[PrintTestReport][60] :x: [Vid Capture]
[PrintTestReport][62] :x: Total item [60], avg [40.8 ms]
[PrintTestReport][60] :x: [Img Xcoding]
[PrintTestReport][62] :x: Total item [60], avg [1.6 ms]
[PrintTestReport][60] :x: [FaceFinder]
[PrintTestReport][62] :x: Total item [60], avg [62.6 ms]
[PrintTestReport][60] :x: [Crop Embedding]
[PrintTestReport][62] :x: Total item [27], avg [95.2 ms]
[PrintTestReport][60] :x: [Classifier]
[PrintTestReport][62] :x: Total item [27], avg [14.4 ms]
[PrintTestReport][60] :x: [Draw Detection]
[PrintTestReport][62] :x: Total item [27], avg [3.0 ms]
[PrintTestReport][60] :x: [EncodingJPEG]
[PrintTestReport][62] :x: Total item [57], avg [64.3 ms]
[PrintTestReport][60] :x: [SendImg_TCP]
[PrintTestReport][62] :x: Total item [54], avg [95.1 ms]

마지막 2개가 별도 thread 에서 돌고 있네요...